Authorisation,HTML,functions
1) Made and authorisation function, that understands if it is a user or admin.
2) Created html pages for admin and reader, where every available function for specific user
3)Implemented authorisation functio.
PS: Made it os that Admin is not and instance of reader, because the admins job is to work with the system, not read books. If admin wants to take boks, then he has to make a seperate account. I made it so that it is easier to understand who is a reader and whos is admin for authorisation, because the system always saw admin as a reader first.